backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
motor==3.3.2
pymongo==4.6.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for API calls)
httpx==0.26.0

# AI/ML
google-generativeai==0.3.2

# File upload support
python-multipart==0.0.6